+++ src/views/home.js
@@ -4,49 +4,12 @@ import { jsx } from "@emotion/core";
 import DiscussionList from "../components/discussion-list";
 import DiscussionForm from "../components/discussion-form";
 import { DialogOverlay, DialogContent } from "@reach/dialog";
+import { useDiscussion } from "../contexts/discussions";
 import "@reach/dialog/styles.css";
 
-const initialDiscussion = JSON.parse(
-  localStorage.getItem("discussions") ||
-    JSON.stringify([
-      {
-        id: new Date("2019-06-13T21:50:18.315Z").getTime(),
-        title: "How kill to Christopher",
-        body: "Some body in the sky",
-        author: "condef5",
-        date: "2019-06-13T21:50:18.315Z"
-      },
-      {
-        id: new Date("2019-06-13T22:50:18.315Z").getTime(),
-        title: "How to build the future today",
-        body: "Some body in the sky",
-        author: "condef5",
-        date: "2019-06-13T22:50:18.315Z"
-      },
-      {
-        id: new Date("2019-06-13T23:50:18.315Z").getTime(),
-        title: "Learn react like a professional",
-        body: "Some body in the sky",
-        author: "condef5",
-        date: "2019-06-13T23:50:18.315Z"
-      },
-      {
-        id: new Date("2019-06-13T13:50:18.315Z").getTime(),
-        title: "The best reply in the history",
-        body: "Some body in the sky",
-        author: "condef5",
-        date: "2019-06-13T13:50:18.315Z"
-      }
-    ])
-);
-
 function Home() {
-  const [discussions, setDiscussions] = React.useState(initialDiscussion);
+  const { discussions, setDiscussions } = useDiscussion();
   const [isDialogOpen, setIsDialogOpen] = React.useState(false);
-
-  React.useEffect(() => {
-    localStorage.setItem("discussions", JSON.stringify(discussions));
-  }, [discussions]);
 
   function handleOpenModal(event) {
     setIsDialogOpen(true);
@@ -67,6 +30,7 @@ function Home() {
         handleCreateDisccusion={handleCreateDisccusion}
       />
       <button
+        aria-label={"Add new discussion"}
         css={{
           position: "fixed",
           bottom: "25px",
